requirements.txt
fastapi==0.116.1
google-cloud-storage==3.4.1
openai-agents==0.13.6
pydantic==2.12.5
uvicorn[standard]==0.35.0
